BpkComposeTest.kt update.
