+++ app/src/androidTest/java/net/skyscanner/backpack/compose/BpkTextTest.kt
@@ -2,11 +2,12 @@ package net.skyscanner.backpack.compose
 
 import androidx.compose.material.Text
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import net.skyscanner.backpack.BpkComposeTest
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class BpkTextTest : BpkComposeTest {
+class BpkTextTest : BpkComposeTest() {
 
   @Test
   fun default() = composable {
